requirements.txt
python-dotenv
pytubefix
openai-whisper
langdetect 
tqdm>=4.65.0 